Fix volatility of conversion functions
Function `uuid_v7_to_timestamptz` can be immutable since given an UUID
it will return the same value for eternity regardless of timezone (the
returned values will compare equal regardless of the timezone).

Function `uuid_timestamptz_to_v7` can be immutable since given a
timestamp with a timezone it is ok to return the same random "filler"
for each invocation since the random value cannot be reliabily used for
anything and there is no requirement that each invocation have to
return a different value.

This commit also adds a `uuid_timestamptz_to_v7_stable` for those cases
where a stable value is required, that is, when executions in different
queries have to return a different value.